Complete refactoring of stores
- file, http, memory, ftp, s3, http_erddap stores are moved to the "implemenations" submodule
- store prototype moved to spec
- all index related files are moved to the "index" submodule
