+++ argopy/stores/implementations/__init__.py
@@ -1,0 +1,11 @@
+#!/bin/env python
+# -*coding: UTF-8 -*-
+#
+# HELP
+#
+# Created by gmaze on 19/12/2024
+__author__ = 'gmaze@ifremer.fr'
+
+import os
+import sys
+import xarray as xr
