+++ argopy/stores/index/argo_index_proto_s3.py
@@ -1,20 +1,25 @@
-"""
-Additional index support for S3
-"""
 import io
 import pandas as pd
 import logging
 from decorator import decorator
 import warnings
 
-from ..utils.checkers import check_index_cols, check_wmo, check_cyc, is_list_of_strings, has_aws_credentials, HAS_BOTO3
-from ..utils.format import redact
-from ..stores import s3store
+from ...utils.checkers import (
+    check_index_cols,
+    check_wmo,
+    check_cyc,
+    is_list_of_strings,
+    has_aws_credentials,
+    HAS_BOTO3,
+)
+from ...utils.format import redact
+from .. import s3store
 
 
 try:
     import pyarrow.csv as csv  # noqa: F401
     import pyarrow as pa
+
     HAS_PYARROW = True
 except ModuleNotFoundError:
     HAS_PYARROW = False
@@ -23,6 +28,7 @@ except ModuleNotFoundError:
         @property
         def Table(self):
             pass
+
     pass
 
 if HAS_BOTO3:
@@ -37,7 +43,9 @@ log = logging.getLogger("argopy.stores.index.s3")
 @decorator
 def requires_pyarrow(func, *args, **kwargs):
     if not HAS_PYARROW:
-        warnings.warn("The 'pyarrow' library is not installed. Please install it to ensure full functionality.")
+        warnings.warn(
+            "The 'pyarrow' library is not installed. Please install it to ensure full functionality."
+        )
     return func(*args, **kwargs)
 
 
@@ -73,6 +81,7 @@ class s3index:
     idx.stats # Data processing stats
 
     """
+
     bucket_name = "argo-gdac-sandbox"
     """Name of the S3 bucket"""
 
@@ -81,29 +90,36 @@ class s3index:
         if has_aws_credentials():
             self.fs = boto3.client("s3")
             try:
-                access_key = self.fs._request_signer._credentials.get_frozen_credentials().access_key
-                log.debug("Found AWS Credentials for access_key='%s'" % redact(access_key, 4))
+                access_key = (
+                    self.fs._request_signer._credentials.get_frozen_credentials().access_key
+                )
+                log.debug(
+                    "Found AWS Credentials for access_key='%s'" % redact(access_key, 4)
+                )
             except:  # noqa: E722
                 pass
         else:
-            self.fs = boto3.client('s3', config=Config(signature_version=UNSIGNED))
-            log.debug("No AWS Credentials found, running UNSIGNED anonymous boto3 requests")
+            self.fs = boto3.client("s3", config=Config(signature_version=UNSIGNED))
+            log.debug(
+                "No AWS Credentials found, running UNSIGNED anonymous boto3 requests"
+            )
             # search engines won't be available !
 
         self.stats_last = {}
         self.stats = {}
-        self._sql_logic = 'unique'
+        self._sql_logic = "unique"
 
     def _sio2pd(self, obj_io: io.StringIO) -> pd.DataFrame:
         def _pd(input_io):
-            this_table = pd.read_csv(input_io,
-                                     sep=",",
-                                     index_col=None,
-                                     header=None,
-                                     skiprows=0,
-                                     nrows=None,
-                                     names=self.colNames
-                                     )
+            this_table = pd.read_csv(
+                input_io,
+                sep=",",
+                index_col=None,
+                header=None,
+                skiprows=0,
+                nrows=None,
+                names=self.colNames,
+            )
             return this_table
 
         index = _pd(obj_io)
@@ -116,14 +132,13 @@ class s3index:
 
     @requires_pyarrow
     def _sio2pq(self, obj_io: io.StringIO) -> pa.Table:
-        input_bytes = io.BytesIO(obj_io.read().encode('utf8'))
+        input_bytes = io.BytesIO(obj_io.read().encode("utf8"))
         if input_bytes.getbuffer().nbytes > 0:
             index = csv.read_csv(
                 input_bytes,
-                read_options=csv.ReadOptions(use_threads=True,
-                                             skip_rows=0,
-                                             column_names=self.colNames
-                                             ),
+                read_options=csv.ReadOptions(
+                    use_threads=True, skip_rows=0, column_names=self.colNames
+                ),
                 convert_options=csv.ConvertOptions(
                     column_types={
                         "date": pa.timestamp("s"),  # , tz="utc"
@@ -149,15 +164,18 @@ class s3index:
                 Key=self.key,
                 ExpressionType="SQL",
                 Expression=sql_expression,
-                InputSerialization={"CSV": {"FileHeaderInfo": "IGNORE",
-                                            'Comments': '#',
-                                            'QuoteEscapeCharacter': '"',
-                                            'RecordDelimiter': '\n',
-                                            'FieldDelimiter': ',',
-                                            'QuoteCharacter': '"',
-                                            'AllowQuotedRecordDelimiter': False
-                                            },
-                                    "CompressionType": self.CompressionType},
+                InputSerialization={
+                    "CSV": {
+                        "FileHeaderInfo": "IGNORE",
+                        "Comments": "#",
+                        "QuoteEscapeCharacter": '"',
+                        "RecordDelimiter": "\n",
+                        "FieldDelimiter": ",",
+                        "QuoteCharacter": '"',
+                        "AllowQuotedRecordDelimiter": False,
+                    },
+                    "CompressionType": self.CompressionType,
+                },
                 OutputSerialization={"CSV": {}},
             )
         except:  # noqa: E722
@@ -169,13 +187,13 @@ class s3index:
         for event in s3_object["Payload"]:
             if "Records" in event:
                 records.append(event["Records"]["Payload"].decode("utf-8"))
-            elif 'Stats' in event:
-                stats = event['Stats']['Details']
+            elif "Stats" in event:
+                stats = event["Stats"]["Details"]
 
         self.stats_last = stats
         self.stats.update({sql_expression: stats})
 
-        return ''.join(r for r in records)
+        return "".join(r for r in records)
 
     def run(self):
         if not is_list_of_strings(self.sql_expression):
@@ -190,14 +208,20 @@ class s3index:
     def search_wmo(self, WMOs, nrows=None):
         WMOs = check_wmo(WMOs)
 
-        if self._sql_logic == 'split':
+        if self._sql_logic == "split":
             sql = []
             for wmo in WMOs:
-                sql.append("SELECT * FROM s3object s WHERE s._1 LIKE '%/{wmo}/profiles/%'".format(wmo=wmo))
+                sql.append(
+                    "SELECT * FROM s3object s WHERE s._1 LIKE '%/{wmo}/profiles/%'".format(
+                        wmo=wmo
+                    )
+                )
         else:
             sql = "SELECT * FROM s3object s WHERE "
             if len(WMOs) > 1:
-                sql += " OR ".join(["s._1 LIKE '%/{wmo}/profiles/%'".format(wmo=wmo) for wmo in WMOs])
+                sql += " OR ".join(
+                    ["s._1 LIKE '%/{wmo}/profiles/%'".format(wmo=wmo) for wmo in WMOs]
+                )
             else:
                 sql += "s._1 LIKE '%/{wmo}/profiles/%'".format(wmo=WMOs[0])
 
@@ -211,14 +235,23 @@ class s3index:
     def search_cyc(self, CYCs, nrows=None):
         CYCs = check_cyc(CYCs)
 
-        if self._sql_logic == 'split':
+        if self._sql_logic == "split":
             sql = []
             for cyc in CYCs:
-                sql.append("SELECT * FROM s3object s WHERE s._1 LIKE '%/%/profiles/%_{cyc:03d}.nc'".format(cyc=cyc))
+                sql.append(
+                    "SELECT * FROM s3object s WHERE s._1 LIKE '%/%/profiles/%_{cyc:03d}.nc'".format(
+                        cyc=cyc
+                    )
+                )
         else:
             sql = "SELECT * FROM s3object s WHERE "
             if len(CYCs) > 1:
-                sql += " OR ".join(["s._1 LIKE '%/%/profiles/%_{cyc:03d}.nc'".format(cyc=cyc) for cyc in CYCs])
+                sql += " OR ".join(
+                    [
+                        "s._1 LIKE '%/%/profiles/%_{cyc:03d}.nc'".format(cyc=cyc)
+                        for cyc in CYCs
+                    ]
+                )
             else:
                 sql += "s._1 LIKE '%/%/profiles/%_{cyc:03d}.nc'".format(cyc=CYCs[0])
 
@@ -233,30 +266,51 @@ class s3index:
         WMOs = check_wmo(WMOs)
         CYCs = check_cyc(CYCs)
 
-        if self._sql_logic == 'split':
+        if self._sql_logic == "split":
             sql = []
             for wmo in WMOs:
                 for cyc in CYCs:
                     sql.append(
-                        "SELECT * FROM s3object s WHERE s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(wmo=wmo,
-                                                                                                            cyc=cyc))
+                        "SELECT * FROM s3object s WHERE s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(
+                            wmo=wmo, cyc=cyc
+                        )
+                    )
         else:
             sql = "SELECT * FROM s3object s WHERE "
             if len(WMOs) > 1:
                 if len(CYCs) > 1:
                     sql += " OR ".join(
-                        ["s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(wmo=wmo, cyc=cyc) for wmo in WMOs for cyc
-                         in
-                         CYCs])
+                        [
+                            "s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(
+                                wmo=wmo, cyc=cyc
+                            )
+                            for wmo in WMOs
+                            for cyc in CYCs
+                        ]
+                    )
                 else:
                     sql += " OR ".join(
-                        ["s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(wmo=wmo, cyc=CYCs[0]) for wmo in WMOs])
+                        [
+                            "s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(
+                                wmo=wmo, cyc=CYCs[0]
+                            )
+                            for wmo in WMOs
+                        ]
+                    )
             else:
                 if len(CYCs) > 1:
                     sql += " OR ".join(
-                        ["s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(wmo=WMOs[0], cyc=cyc) for cyc in CYCs])
+                        [
+                            "s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(
+                                wmo=WMOs[0], cyc=cyc
+                            )
+                            for cyc in CYCs
+                        ]
+                    )
                 else:
-                    sql += "s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(wmo=WMOs[0], cyc=CYCs[0])
+                    sql += "s._1 LIKE '%/{wmo}/profiles/%_{cyc:03d}.nc'".format(
+                        wmo=WMOs[0], cyc=CYCs[0]
+                    )
 
         if nrows is not None:
             sql += " LIMIT %i" % nrows
@@ -268,8 +322,8 @@ class s3index:
     @property
     def pd(self) -> pd.DataFrame:
         """Return search result as a :class:`pd.DataFrame`"""
-        if not hasattr(self, 'search'):
-            raise Exception('Execute a search first !')
+        if not hasattr(self, "search"):
+            raise Exception("Execute a search first !")
         else:
             return self._sio2pd(self.search)
 
@@ -277,8 +331,8 @@ class s3index:
     @requires_pyarrow
     def pq(self) -> pa.Table:
         """Return search result as a :class:`pa.Table`"""
-        if not hasattr(self, 'search'):
-            raise Exception('Execute a search first !')
+        if not hasattr(self, "search"):
+            raise Exception("Execute a search first !")
         else:
             return self._sio2pq(self.search)
 
@@ -309,25 +363,30 @@ class s3index_core(s3index):
     @property
     @requires_pyarrow
     def empty_pq(self):
-        return pa.Table.from_pydict({'file': [],
-                                     'date': [],
-                                     'latitude': [],
-                                     'longitude': [],
-                                     'ocean': [],
-                                     'profiler_type': [],
-                                     'institution': [],
-                                     'date_update': [],
-                                     },
-                                    schema=pa.schema([
-                                        ("file", pa.string()),
-                                        ("date", pa.timestamp('s')),
-                                        ("latitude", pa.float64()),
-                                        ("longitude", pa.float64()),
-                                        ("ocean", pa.string()),
-                                        ("profiler_type", pa.int64()),
-                                        ("institution", pa.string()),
-                                        ("date_update", pa.timestamp('s')),
-                                    ]))
+        return pa.Table.from_pydict(
+            {
+                "file": [],
+                "date": [],
+                "latitude": [],
+                "longitude": [],
+                "ocean": [],
+                "profiler_type": [],
+                "institution": [],
+                "date_update": [],
+            },
+            schema=pa.schema(
+                [
+                    ("file", pa.string()),
+                    ("date", pa.timestamp("s")),
+                    ("latitude", pa.float64()),
+                    ("longitude", pa.float64()),
+                    ("ocean", pa.string()),
+                    ("profiler_type", pa.int64()),
+                    ("institution", pa.string()),
+                    ("date_update", pa.timestamp("s")),
+                ]
+            ),
+        )
 
 
 class s3index_bgc_bio(s3index):
@@ -357,29 +416,34 @@ class s3index_bgc_bio(s3index):
     @property
     @requires_pyarrow
     def empty_pq(self):
-        return pa.Table.from_pydict({'file': [],
-                                     'date': [],
-                                     'latitude': [],
-                                     'longitude': [],
-                                     'ocean': [],
-                                     'profiler_type': [],
-                                     'institution': [],
-                                     "parameters": [],
-                                     "parameter_data_mode": [],
-                                     'date_update': [],
-                                     },
-                                    schema=pa.schema([
-                                        ("file", pa.string()),
-                                        ("date", pa.timestamp('s')),
-                                        ("latitude", pa.float64()),
-                                        ("longitude", pa.float64()),
-                                        ("ocean", pa.string()),
-                                        ("profiler_type", pa.int64()),
-                                        ("institution", pa.string()),
-                                        ("parameters", pa.string()),
-                                        ("parameter_data_mode", pa.string()),
-                                        ("date_update", pa.timestamp('s')),
-                                    ]))
+        return pa.Table.from_pydict(
+            {
+                "file": [],
+                "date": [],
+                "latitude": [],
+                "longitude": [],
+                "ocean": [],
+                "profiler_type": [],
+                "institution": [],
+                "parameters": [],
+                "parameter_data_mode": [],
+                "date_update": [],
+            },
+            schema=pa.schema(
+                [
+                    ("file", pa.string()),
+                    ("date", pa.timestamp("s")),
+                    ("latitude", pa.float64()),
+                    ("longitude", pa.float64()),
+                    ("ocean", pa.string()),
+                    ("profiler_type", pa.int64()),
+                    ("institution", pa.string()),
+                    ("parameters", pa.string()),
+                    ("parameter_data_mode", pa.string()),
+                    ("date_update", pa.timestamp("s")),
+                ]
+            ),
+        )
 
 
 class s3index_bgc_synthetic(s3index_bgc_bio):
@@ -414,33 +478,45 @@ def search_s3(func, *args, **kwargs):
     """
     idx = args[0]
 
-    if func.__name__ == 'search_wmo' and not hasattr(idx, 'index') and isinstance(idx.fs['src'], s3store):
+    if (
+        func.__name__ == "search_wmo"
+        and not hasattr(idx, "index")
+        and isinstance(idx.fs["src"], s3store)
+    ):
         WMOs, nrows = args[1], args[2]
         WMOs = check_wmo(WMOs)  # Check and return a valid list of WMOs
         log.debug(
             "Argo index searching for WMOs=[%s] using boto3 SQL request ..."
             % ";".join([str(wmo) for wmo in WMOs])
         )
-        idx.fs['s3'].search_wmo(WMOs, nrows=nrows)
+        idx.fs["s3"].search_wmo(WMOs, nrows=nrows)
         idx.search_type = {"WMO": WMOs}
-        idx.search_filter = idx.fs['s3'].sql_expression
-        idx.search = getattr(idx.fs['s3'], idx.ext)
+        idx.search_filter = idx.fs["s3"].sql_expression
+        idx.search = getattr(idx.fs["s3"], idx.ext)
         return idx
 
-    if func.__name__ == 'search_cyc' and not hasattr(idx, 'index') and isinstance(idx.fs['src'], s3store):
+    if (
+        func.__name__ == "search_cyc"
+        and not hasattr(idx, "index")
+        and isinstance(idx.fs["src"], s3store)
+    ):
         CYCs, nrows = args[1], args[2]
         CYCs = check_cyc(CYCs)  # Check and return a valid list of CYCs
         log.debug(
             "Argo index searching for CYCs=[%s] using boto3 SQL request ..."
             % (";".join([str(cyc) for cyc in CYCs]))
         )
-        idx.fs['s3'].search_cyc(CYCs, nrows=nrows)
+        idx.fs["s3"].search_cyc(CYCs, nrows=nrows)
         idx.search_type = {"CYC": CYCs}
-        idx.search_filter = idx.fs['s3'].sql_expression
-        idx.search = getattr(idx.fs['s3'], idx.ext)
+        idx.search_filter = idx.fs["s3"].sql_expression
+        idx.search = getattr(idx.fs["s3"], idx.ext)
         return idx
 
-    if func.__name__ == 'search_wmo_cyc' and not hasattr(idx, 'index') and isinstance(idx.fs['src'], s3store):
+    if (
+        func.__name__ == "search_wmo_cyc"
+        and not hasattr(idx, "index")
+        and isinstance(idx.fs["src"], s3store)
+    ):
         WMOs, CYCs, nrows = args[1], args[2], args[3]
         WMOs = check_wmo(WMOs)  # Check and return a valid list of WMOs
         CYCs = check_cyc(CYCs)  # Check and return a valid list of CYCs
@@ -451,10 +527,10 @@ def search_s3(func, *args, **kwargs):
                 ";".join([str(cyc) for cyc in CYCs]),
             )
         )
-        idx.fs['s3'].search_wmo_cyc(WMOs, CYCs, nrows=nrows)
+        idx.fs["s3"].search_wmo_cyc(WMOs, CYCs, nrows=nrows)
         idx.search_type = {"WMO": WMOs, "CYC": CYCs}
-        idx.search_filter = idx.fs['s3'].sql_expression
-        idx.search = getattr(idx.fs['s3'], idx.ext)
+        idx.search_filter = idx.fs["s3"].sql_expression
+        idx.search = getattr(idx.fs["s3"], idx.ext)
         return idx
 
     return func(*args, **kwargs)
